RSATools/requirements.txt
requests>=2.32.3
gmpy2>=2.2.1
pycryptodome>=3.21.0